A heap of changes to fix startup-to-GUI
This change replaces the network port mapping with network namespacing,
mainly because the Venus system depends on accessing the legacy GUI process
via port 81.

The Venusian now sets up a local bridged network and attaches the Venus
instance(s) to it. The network address is fixed to 192.168.42.0/24
(for now).

Also, there is a new setup option "--ext=IPADDR" which also attaches
the Venus instance to the given external IP address. Use "--ext=dyn" to
request an address via DHCP.

The web frontend now works, both in GUIv2 and GUIv1 mode, with the
caveats that (a) authorization is turned off (write an empty string
to /var/lib/venusian/conf/vncpassword.txt), (b) *not* using an
external address is untested.